--- pyIOFlash.sln
@@ -1,22 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28803.156
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "pyIOFlash", "pyIOFlash.pyproj", "{40AC6589-0E5E-46A3-8E63-DA60E6D982A9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{40AC6589-0E5E-46A3-8E63-DA60E6D982A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40AC6589-0E5E-46A3-8E63-DA60E6D982A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6FA77E7A-3A12-4062-A0FA-C8D98B196C90}
-	EndGlobalSection
-EndGlobal
