+++ pyioflash/__init__.py
@@ -4,7 +4,10 @@ and process FLASH4 HDF5 plt and chk files
 """
 
 # Used to import and handle simulation data
-from .pyio import SimulationData
+from pyioflash.simulation import SimulationData
+from pyioflash.simulation import NameData
 
 # Used to create professional 2D plots
-from .pyio_plot import SimulationPlot
+from pyioflash.visual import SimulationPlot
+from pyioflash.visual import FigureOptions, PlotOptions, AnimationOptions
+from pyioflash.visual import Plane
