+++ pyioflash/simulation/__init__.py
@@ -1,0 +1,11 @@
+"""
+Python Module for providing methods to import
+and process FLASH4 HDF5 plt and chk files
+"""
+
+# Used to import and handle simulation data
+from pyioflash.simulation.data import SimulationData
+
+# Used to assist in defining time series data
+from pyioflash.simulation.series import NameData
+
